Simplified domain package structure
