+++ src/test/groovy/game/SavedGameSpec.groovy
@@ -1,0 +1,11 @@
+package game
+
+import spock.lang.Specification
+
+class SavedGameSpec extends Specification {
+    void setup() {
+    }
+
+    void cleanup() {
+    }
+}
